+++ Thrower.NuGetPackager/CHANGELOG.md
@@ -1,4 +1,8 @@
-﻿### v2.0.2 (2015-10-24) ###
+﻿### v2.0.3 (2015-10-31) ###
+* FastMember.ObjectAccessor now implements IDictionary&lt;string, object&gt;.
+* Added HttpException for Web API projects.
+
+### v2.0.2 (2015-10-24) ###
 * Embedded FastMember.
 * Added a simple reflection-based object validator.